% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{make_object}
\alias{make_object}
\title{make_object 関数}
\usage{
make_object(y_range, vec_x, vec_y, x_range = c(0, 1))
}
\arguments{
\item{y_range}{y軸の正規化範囲を指定します。}

\item{vec_x}{x軸に使用するベクトル}

\item{vec_y}{y軸に使用するベクトル}

\item{x_range}{x軸の正規化範囲を指定します。デフォルトは [0, 1] です。}
}
\value{
グラフィカルなオブジェクトを含むリスト
}
\description{
この関数は、グラフィカルなオブジェクトを生成するためのユーティリティ関数です。指定した範囲にベクトルを正規化し、x軸やy軸の操作をサポートします。
}
\examples{
# 例: グラフィカルオブジェクトを生成
obj <- make_object(y_range = c(0, 100), vec_x = 1:10, vec_y = c(20, 30, 40, 50, 60, 70, 80, 90, 100, 110))

# グラフィカルオブジェクトの使用例
obj$xaxis()
obj$lines()
obj$points()

}
