% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{plane_plot}
\alias{plane_plot}
\title{plane_plot 関数}
\usage{
plane_plot(...)
}
\arguments{
\item{...}{他のプロット関数に渡す追加の引数}
}
\value{
空のプロット領域
}
\description{
この関数は、空のプロット領域を作成します。デフォルトでは、x軸とy軸の範囲は[0, 1]に設定され、ラベルや軸は表示されません。
}
\examples{
# 例: 空のプロット領域を作成
plane_plot()

# 他のプロット関数と一緒に使用する例
plane_plot(xlim = c(0, 10), ylim = c(0, 5), xlab = "X軸", ylab = "Y軸")
}
